+++ agreement_operating_unit/readme/DESCRIPTION.rst
@@ -1,1 +1,1 @@
-This module adds operating unit information to agreements and service profiles.
+This module adds operating unit information to agreements
