+++ agreement_serviceprofile_operating_unit/__init__.py
@@ -1,2 +1,2 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-from . import agreement
+from . import models
