+++ agreement_serviceprofile_operating_unit/i18n/pt.po
@@ -1,0 +1,51 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* agreement_operating_unit
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"PO-Revision-Date: 2020-02-29 15:13+0000\n"
+"Last-Translator: Pedro Castro Silva <pedrocs@exo.pt>\n"
+"Language-Team: none\n"
+"Language: pt\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=2; plural=n > 1;\n"
+"X-Generator: Weblate 3.10\n"
+
+#. module: agreement_operating_unit
+#: model:ir.model,name:agreement_operating_unit.model_agreement
+msgid "Agreement"
+msgstr "Acordo"
+
+#. module: agreement_operating_unit
+#: model:ir.model,name:agreement_operating_unit.model_agreement_serviceprofile
+msgid "Agreement Service Profiles"
+msgstr "Perfis de Serviços do Acordo"
+
+#. module: agreement_operating_unit
+#: model:ir.model.fields,field_description:agreement_operating_unit.field_agreement__display_name
+#: model:ir.model.fields,field_description:agreement_operating_unit.field_agreement_serviceprofile__display_name
+msgid "Display Name"
+msgstr ""
+
+#. module: agreement_operating_unit
+#: model:ir.model.fields,field_description:agreement_operating_unit.field_agreement__id
+#: model:ir.model.fields,field_description:agreement_operating_unit.field_agreement_serviceprofile__id
+msgid "ID"
+msgstr ""
+
+#. module: agreement_operating_unit
+#: model:ir.model.fields,field_description:agreement_operating_unit.field_agreement____last_update
+#: model:ir.model.fields,field_description:agreement_operating_unit.field_agreement_serviceprofile____last_update
+msgid "Last Modified on"
+msgstr ""
+
+#. module: agreement_operating_unit
+#: model:ir.model.fields,field_description:agreement_operating_unit.field_agreement__operating_unit_id
+#: model:ir.model.fields,field_description:agreement_operating_unit.field_agreement_serviceprofile__operating_unit_id
+msgid "Operating Unit"
+msgstr "Unidade Operacional"
