+++ agreement_serviceprofile_operating_unit/models/__init__.py
@@ -1,2 +1,2 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-from . import agreement
+from . import agreement_serviceprofile
