+++ agreement_serviceprofile_operating_unit/readme/CREDITS.rst
@@ -1,0 +1,2 @@
+* Open Source Integrators
+* Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
