+++ agreement_serviceprofile_operating_unit/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+This module adds operating unit information to agreement service profiles.
