+++ agreement_serviceprofile_operating_unit/security/aggreement_security.xml
@@ -4,13 +4,16 @@
     Copyright (C) 2021 Serpent Consulting Services
     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
 
-    <record id="ir_rule_agreement_operating_unit" model="ir.rule">
-        <field name="model_id" ref="agreement.model_agreement" />
+    <record id="ir_rule_agreement_serviceprofile_operating_unit" model="ir.rule">
+        <field
+            name="model_id"
+            ref="agreement_serviceprofile.model_agreement_serviceprofile"
+        />
         <field
             name="domain_force"
         >['|', ('operating_unit_id','=',False), ('operating_unit_id','in',user.operating_unit_ids.ids)]
         </field>
-        <field name="name">Agreement Operating Unit</field>
+        <field name="name">Agreement Serviceprofile Operating Unit</field>
         <field name="global" eval="True" />
         <field eval="0" name="perm_unlink" />
         <field eval="0" name="perm_write" />
