+++ setup/agreement_serviceprofile_operating_unit/odoo/addons/agreement_serviceprofile_operating_unit
@@ -1,0 +1,1 @@
+../../../../agreement_serviceprofile_operating_unit
